terraform: handle count fields for data sources
